Update: refactoring and add !ask command
