+++ package.json
@@ -2,9 +2,9 @@
   "name": "api-openai-discord-channel",
   "version": "1.0.0",
   "description": "Discord bot with open ai integration",
-  "main": "index.js",
+  "main": "src/index.js",
   "scripts": {
-    "start": "node index.js",
+    "start": "node src/index.js",
     "dev": "nodemon index.js"
   },
   "repository": {
