+++ src/index.js
@@ -1,7 +1,7 @@
 // Importamos las dependencias necesarias
 require('dotenv').config();
 const { Client, GatewayIntentBits } = require('discord.js');
-const { setupOpenAI, askOpenAI } = require('./ai.js');
+const askCommand = require('./commands/ask.js');
 
 // Configuramos los intents para el cliente de Discord
 const client = new Client({ intents: [
@@ -9,8 +9,6 @@ const client = new Client({ intents: [
     GatewayIntentBits.GuildMessages,
     GatewayIntentBits.MessageContent
 ]})
-
-const openai = setupOpenAI();
 
 // Define el ID del canal en el que el bot debe responder, como para un layer extra más allá de privilegios en el canal
 const targetChannelId = process.env.DISCORD_CHANNEL_ID;
@@ -20,28 +18,8 @@ client.on('messageCreate', async (discordMessage) => {
     // Ignoramos mensajes de bots y mensajes fuera del canal objetivo
     if (discordMessage.author.bot || discordMessage.channel.id !== targetChannelId) return;
 
-    try {
-        // Creamos una conversación inicial
-        const message = [
-            {
-                role: "user",
-                content: discordMessage.content
-            }
-        ];
-
-        // Enviamos la consulta a la API de OpenAI
-        const responseMessage = await askOpenAI(openai, discordMessage, message)
-
-        discordMessage.reply(responseMessage);
-    } catch (err) {
-        // Registramos el error en caso de que ocurra alguno
-        console.error('Error al procesar el mensaje:', err);
-
-        // Imprimimos la respuesta de error de la API de OpenAI
-        if (err.response && err.response.data) {
-            console.error('Error en la respuesta de OpenAI:', err.response.data);
-        }
-    }
+    // Comprobamos si el mensaje comienza con el prefijo '!ask' y ejecutamos el comando
+    if (discordMessage.content.startsWith('!ask')) await askCommand(discordMessage);
 });
 
 // Iniciamos sesión en Discord y mostramos un mensaje cuando el bot esté listo
